Extract waku initialisation in separate component
